move user.yaml to user.common
we need to be able to provide hard-coded user data
for our automated tests. And on occasion that data
needs to be specified per scenario (b/c different
data overrides are specific to certain scenarios).

This commit modifies the existing user data to
better accomplish this.

1. Adds user.common and user.%{scenario} to the
hierarchy. Allowing a place for common and
scenario specific user data for testing.

2. moves user.yaml to user.common. This assumes that
users can supply fresh user specific overrides
to user.yaml, and that user.common show examples of
things that a user may want to override. It is possible
that this will have to change and that
user.common should really be user.2_role