Update README and separate master setup from node
If you run the same process on the master and non-master
nodes, then puppet-agent runs fail.  This may be a bug
in the node_terminus module, but this is also not necessarily
needed on non-master nodes.

In addition the single setup script was install puppetmaster
code on every node, which isn't necessarily appropriate either.

Still TBD:
 - add puppet agent to run automatically on non build/master
 nodes